+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/services/plugins/ThirdPartyPluginService.java
@@ -101,9 +101,6 @@ public class ThirdPartyPluginService {
         // mcMMO Integration
         if (isPluginInstalled("mcMMO")) {
             try {
-                // This makes sure that the FakeEvent interface is present.
-                // Class.forName("com.gmail.nossr50.events.fake.FakeEvent");
-
                 new McMMOIntegration(plugin);
                 isMcMMOInstalled = true;
             } catch (Exception | LinkageError x) {
@@ -170,7 +167,7 @@ public class ThirdPartyPluginService {
      * @return Whether this is a fake event
      */
     public boolean isEventFaked(@Nonnull Event event) {
-        // TODO: Change this to FakeEvent once the new mcMMO build was released
+        // This can be changed to "FakeEvent" in a later version
         return isMcMMOInstalled && event instanceof FakeBlockBreakEvent;
     }