+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/settings/TalismanEnchantment.java
@@ -21,7 +21,7 @@ public class TalismanEnchantment extends ItemSetting<Boolean> {
     private final int level;
 
     public TalismanEnchantment(@Nonnull Enchantment enchantment, int level) {
-        super("allow-enchantments." + enchantment.getKey().getKey() + ".level." + level, true);
+        super("allow-enchantments." + enchantment.getKey().getNamespace() + '.' + enchantment.getKey().getKey() + ".level." + level, true);
 
         this.enchantment = enchantment;
         this.level = level;