+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 ## Release Candidate 18 (TBD)
 
 #### Additions
+* The Smelters Pick now also works on Ancient Debris
 
 #### Changes
 * Removed 1.13 support
@@ -48,6 +49,11 @@
 * Fixed #2469
 * Fixed #2509
 * Fixed #2499
+* Fixed #2527
+* Fixed #2519
+* Fixed #2517
+* Fixed Magician Talisman sometimes drawing invalid enchantments
+* Fixed id conflicts for external Enchantment sources (e.g. plugins) for the Magician Talisman settings
 
 ## Release Candidate 17 (17 Oct 2020)